refactor: state 패키지를 domain 아래로 이동
