+++ src/main/java/com/finnect/user/adapter/in/security/AuthenticationFilter.java
@@ -7,7 +7,7 @@ import com.finnect.user.application.port.in.GetPersonalNameQuery;
 import com.finnect.user.application.port.in.IssueUseCase;
 import com.finnect.user.application.port.in.command.IssueCommand;
 import com.finnect.user.domain.UserAuthentication;
-import com.finnect.user.state.TokenPairState;
+import com.finnect.user.domain.state.TokenPairState;
 import com.finnect.common.vo.UserId;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
