+++ src/main/java/com/finnect/user/adapter/in/security/AuthenticationProviderImpl.java
@@ -3,19 +3,14 @@ package com.finnect.user.adapter.in.security;
 import com.finnect.user.adapter.in.security.util.AuthenticationUtils;
 import com.finnect.user.application.port.in.AuthenticateUseCase;
 import com.finnect.user.application.port.in.command.AuthenticateCommand;
-import com.finnect.user.state.UserAuthenticationState;
+import com.finnect.user.domain.state.UserAuthenticationState;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Component;
-
-import java.util.Collection;
-import java.util.HashSet;
 
 @Slf4j
 @Component
