+++ src/main/java/com/finnect/user/application/port/in/AuthorizeUseCase.java
@@ -1,7 +1,7 @@
 package com.finnect.user.application.port.in;
 
 import com.finnect.user.application.port.in.command.AuthorizeCommand;
-import com.finnect.user.state.UserAuthenticationState;
+import com.finnect.user.domain.state.UserAuthenticationState;
 
 public interface AuthorizeUseCase {
