+++ src/main/java/com/finnect/user/application/port/in/ReissueUseCase.java
@@ -2,7 +2,7 @@ package com.finnect.user.application.port.in;
 
 import com.finnect.user.application.port.in.command.ReissueCommand;
 import com.finnect.user.application.port.in.command.ReissueWorkspaceCommand;
-import com.finnect.user.state.AccessTokenState;
+import com.finnect.user.domain.state.AccessTokenState;
 
 public interface ReissueUseCase {
