+++ src/main/java/com/finnect/user/application/service/IssueService.java
@@ -11,8 +11,8 @@ import com.finnect.user.application.port.out.GenerateAccessTokenPort;
 import com.finnect.user.application.port.out.LoadRefreshTokenPort;
 import com.finnect.user.application.port.out.SaveRefreshTokenPort;
 import com.finnect.user.domain.*;
-import com.finnect.user.state.AccessTokenState;
-import com.finnect.user.state.TokenPairState;
+import com.finnect.user.domain.state.AccessTokenState;
+import com.finnect.user.domain.state.TokenPairState;
 import com.finnect.common.vo.UserId;
 import com.finnect.common.vo.WorkspaceAuthority;
 import com.finnect.workspace.application.port.in.CheckWorkspaceQuery;
