+++ src/main/java/com/finnect/user/domain/UserDetailsImpl.java
@@ -1,6 +1,6 @@
 package com.finnect.user.domain;
 
-import com.finnect.user.state.UserState;
+import com.finnect.user.domain.state.UserState;
 import com.finnect.common.vo.UserId;
 import com.finnect.common.vo.WorkspaceAuthority;
 import com.finnect.common.vo.WorkspaceId;
