+++ src/main/java/com/finnect/user/domain/state/EmailCodeState.java
@@ -1,4 +1,4 @@
-package com.finnect.user.state;
+package com.finnect.user.domain.state;
 
 public interface EmailCodeState {
