+++ src/main/java/com/finnect/user/domain/state/UserState.java
@@ -1,4 +1,4 @@
-package com.finnect.user.state;
+package com.finnect.user.domain.state;
 
 import com.finnect.common.vo.UserId;
 import com.finnect.common.vo.WorkspaceId;
